Factored out common setup and teardown into parent class
