+++ src/test/java/tests/GenericTest.java
@@ -1,0 +1,23 @@
+package tests;
+
+import org.openqa.selenium.WebDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeSuite;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+
+
+public class GenericTest {
+    WebDriver driver;
+
+    @BeforeSuite
+    public void initialize() {
+        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
+        driver = new ChromeDriver();
+    }
+
+    @AfterMethod
+    public void  tearDown() {
+        driver.close();
+    }
+}
