+++ src/test/java/tests/Test11FileUpload.java
@@ -1,25 +1,10 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
-public class Test11FileUpload {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test11FileUpload extends GenericTest {
 
     @Test
     public void Test11FileUpload() throws InterruptedException {
@@ -35,8 +20,4 @@ public class Test11FileUpload {
         Assert.assertTrue(driver.findElement(uploadedLabelXpath).isDisplayed());
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
