+++ src/test/java/tests/Test12FloatingMenu.java
@@ -1,26 +1,11 @@
 package tests;
 
 import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
-public class Test12FloatingMenu {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test12FloatingMenu extends GenericTest{
 
     @Test
     public void Test12FloatingMenu() throws InterruptedException {
@@ -37,8 +22,4 @@ public class Test12FloatingMenu {
 
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
