+++ src/test/java/tests/Test13Iframe.java
@@ -1,28 +1,14 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
-public class Test13Iframe {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test13Iframe extends GenericTest {
 
     @Test
-    public void Test13Iframe() throws InterruptedException {
+    public void Test13Iframe() {
         String url = "http://localhost:7080/iframe";
         driver.get(url);
 
@@ -34,11 +20,6 @@ public class Test13Iframe {
 
         //assert that the text is what was typed in previously
         Assert.assertEquals(driver.findElement(By.xpath("//body/p")).getText(),"Hello World!");
-
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
