+++ src/test/java/tests/Test14MouseHover.java
@@ -1,29 +1,13 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.interactions.Actions;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
-
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
-import java.util.ArrayList;
 import java.util.List;
 
-public class Test14MouseHover {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test14MouseHover extends GenericTest {
 
     @Test
     public void Test14MouseHover() {
@@ -51,11 +35,6 @@ public class Test14MouseHover {
             String figureCaptionText = figures.get(i).findElement(By.xpath("descendant::h5")).getText();
             Assert.assertEquals(figureCaptionText,figureCaptionKeys[i]);
         }
-
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
