+++ src/test/java/tests/Test15JavaScriptAlerts.java
@@ -1,28 +1,13 @@
 package tests;
 
 import org.openqa.selenium.Alert;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
-public class Test15JavaScriptAlerts {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test15JavaScriptAlerts extends GenericTest {
 
     @Test
     public void Test15JavaScriptAlerts() {
@@ -59,8 +44,4 @@ public class Test15JavaScriptAlerts {
 
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.quit();
-    }
 }
