+++ src/test/java/tests/Test17OpenInNewTab.java
@@ -1,27 +1,11 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
-
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
 import java.util.ArrayList;
 
-public class Test17OpenInNewTab {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test17OpenInNewTab extends GenericTest {
 
     @Test
     public void Test17OpenInNewTab() {
@@ -41,8 +25,4 @@ public class Test17OpenInNewTab {
         driver.switchTo().defaultContent();
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.quit();
-    }
 }
