+++ src/test/java/tests/Test18NotificationMessage.java
@@ -1,25 +1,10 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
-public class Test18NotificationMessage {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test18NotificationMessage extends GenericTest {
 
     @Test
     public void Test18NotificationMessage() {
@@ -50,8 +35,4 @@ public class Test18NotificationMessage {
         Assert.assertTrue(notificationFlag);
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
