+++ src/test/java/tests/Test1LoginSuccess.java
@@ -1,24 +1,13 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
 import pages.LoginPage;
 import pages.SecureAreaPage;
 
-public class Test1LoginSuccess {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test1LoginSuccess extends GenericTest {
 
     @Test
     public void test1LoginSuccess() {
@@ -40,8 +29,4 @@ public class Test1LoginSuccess {
         Assert.assertTrue(successTextFlag);
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
