+++ src/test/java/tests/Test2LoginFailure.java
@@ -1,23 +1,12 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
 import pages.LoginPage;
 
-public class Test2LoginFailure {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test2LoginFailure extends GenericTest {
 
     @Test
     public void test2LoginFailure() {
@@ -38,8 +27,4 @@ public class Test2LoginFailure {
         Assert.assertTrue(isErrorLabelDisplayed);
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
