+++ src/test/java/tests/Test3CheckBoxes.java
@@ -1,22 +1,11 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-public class Test3CheckBoxes {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test3CheckBoxes extends GenericTest {
 
     @Test
     public void Test3CheckBoxes() {
@@ -42,12 +31,6 @@ public class Test3CheckBoxes {
         Boolean checkBoxTwoResult = checkBoxTwo.isSelected();
         Assert.assertTrue(checkBoxOneResult);
         Assert.assertFalse(checkBoxTwoResult);
-
-
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
