+++ src/test/java/tests/Test5DragAndDrop.java
@@ -1,24 +1,13 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.interactions.Actions;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
 
-public class Test5DragAndDrop {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test5DragAndDrop extends GenericTest {
 
     @Test
     public void Test5DragAndDrop() throws InterruptedException {
@@ -51,8 +40,4 @@ public class Test5DragAndDrop {
         Assert.assertEquals(boxBTextAfter,"A");
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.quit();
-    }
 }
