+++ src/test/java/tests/Test6Dropdown.java
@@ -1,23 +1,12 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.Select;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-public class Test6Dropdown {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test6Dropdown extends GenericTest {
 
     @Test
     public void Test6Dropdown() {
@@ -39,8 +28,4 @@ public class Test6Dropdown {
         Assert.assertEquals(selected.getAttribute("value"),optionTwo);
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
