+++ src/test/java/tests/Test9DynamicLoading.java
@@ -1,27 +1,13 @@
 package tests;
 
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.By;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
 
-import pages.LoginPage;
-import pages.SecureAreaPage;
-
-public class Test9DynamicLoading {
-    WebDriver driver;
-
-    @BeforeSuite
-    public void initialize() {
-        System.setProperty("webdriver.chrome.driver", "src/test/java/drivers/chromedriver");
-        driver = new ChromeDriver();
-    }
+public class Test9DynamicLoading extends GenericTest {
 
     @Test
     public void Test9DynamicLoading() {
@@ -41,8 +27,4 @@ public class Test9DynamicLoading {
         Assert.assertTrue(finishedTitle.isDisplayed());
     }
 
-    @AfterMethod
-    public void  tearDown() {
-        driver.close();
-    }
 }
